+++ dto/src/main/java/com/kairos/user/staff/staff_settings/StaffActivitySettingDTO.java
@@ -12,8 +12,8 @@ public class StaffActivitySettingDTO {
     private BigInteger activityId;
     private Long unitPositionId;
     private Long unitId;
-    private int shortestTime;
-    private int longestTime;
+    private Short shortestTime;
+    private Short longestTime;
     private Integer minLength;
     private Integer maxThisActivityPerShift;
     private boolean eligibleForMove;
@@ -27,8 +27,8 @@ public class StaffActivitySettingDTO {
         //Default Constructor
     }
 
-    public StaffActivitySettingDTO(BigInteger activityId, Long unitPositionId, int shortestTime,
-                                   int longestTime, Integer minLength, Integer maxThisActivityPerShift, boolean eligibleForMove) {
+    public StaffActivitySettingDTO(BigInteger activityId, Long unitPositionId, Short shortestTime,
+                                   Short longestTime, Integer minLength, Integer maxThisActivityPerShift, boolean eligibleForMove) {
         this.activityId = activityId;
         this.unitPositionId = unitPositionId;
         this.shortestTime = shortestTime;
@@ -78,19 +78,19 @@ public class StaffActivitySettingDTO {
         this.unitId = unitId;
     }
 
-    public int getShortestTime() {
+    public Short getShortestTime() {
         return shortestTime;
     }
 
-    public void setShortestTime(int shortestTime) {
+    public void setShortestTime(Short shortestTime) {
         this.shortestTime = shortestTime;
     }
 
-    public int getLongestTime() {
+    public Short getLongestTime() {
         return longestTime;
     }
 
-    public void setLongestTime(int longestTime) {
+    public void setLongestTime(Short longestTime) {
         this.longestTime = longestTime;
     }
